fix: make fork produce a running child that completes and returns cleanly
Resolve the fork/exec/waitpid flow — the shell can now spawn a child
process that runs, prints output, exits, and lets the parent resume.
Previously the child never executed (silent after fork).

**Identity map violation (Bug 1):** `cow_setup_fork` and
`handle_cow_fault` cast physical addresses as pointers in VM's
user-space context. VM has no identity map, so writes corrupt random
VM memory. Fix: move RW-clearing into ring 0
(`vm_paging_fork_x86_64`); rewrite `cow_setup_fork` to use
`vm_mappage`/`vm_unmappage`; replace `pagetable::walk` with
`VM_PAGING_WALK_PAGE` kernel call in `handle_cow_fault`.

**Child inherits SENDING flag (Bug 2):** Parent was in SENDREC at
fork, copying both SENDING and RECEIVING to the child.
`will_receive()` rejects SENDING destinations, so PM's SENDNB to
the child returned ENOTREADY silently. Fix: clear SENDING on child
in `do_fork_handler`.

**COW-shared page double-free:** `handle_exit` called SYS_CLEAR
directly, whose `release_address_space` freed all user pages
including COW-shared pages still in use by the parent. In C MINIX,
PM calls `vm_exit()` first (VM decrements PhysBlock refcounts),
then SYS_CLEAR. Fix:
- Add `VM_EXIT` notification in `handle_exit` before SYS_CLEAR
- Make `vm_destroy` check PhysBlock refcounts before freeing
- Skip user data pages in `release_address_space` (only free
  page table pages)

**`static mut` UB from COW:** `copy_from_slice` on `cmd_path`
(a `static mut [u8; 256]`) panics after fork because the Rust
compiler's `&mut` uniqueness assumption is violated by hardware page
remap. Fix: replace all `copy_from_slice` calls on `static mut`
buffers with manual byte loops.

**Stale Vmproc CR3 after exec:** `vm_get_addrspace` now checks the
kernel's authoritative `p_seg.p_cr3` before falling back to VM's
`vm_pml4_phys`, which can be stale after `exec_replace`.

19 files changed, 807 insertions(+), 545 deletions(-)

+++ crates/servers/src/vfs/pm.rs
@@ -161,7 +161,7 @@ pub fn service_pm() -> i32 {
             #[cfg(target_os = "none")]
             {
                 let mut buf = [0u8; 64];
-                buf[4..8].copy_from_slice(&(VFS_PM_EXEC_REPLY as i32).to_le_bytes());
+                buf[4..8].copy_from_slice(&VFS_PM_EXEC_REPLY.to_le_bytes());
                 buf[8..12].copy_from_slice(&_proc_e.to_le_bytes());
                 buf[12..16].copy_from_slice(&ENOSYS.to_le_bytes()); // STATUS
                 unsafe { send_raw_to_pm(&buf) };
@@ -204,7 +204,7 @@ pub fn service_pm() -> i32 {
             #[cfg(target_os = "none")]
             {
                 let mut buf = [0u8; 64];
-                buf[4..8].copy_from_slice(&(_reply_type as i32).to_le_bytes());
+                buf[4..8].copy_from_slice(&_reply_type.to_le_bytes());
                 buf[8..12].copy_from_slice(&proc_e.to_le_bytes());
                 unsafe { send_raw_to_pm(&buf) };
             }
